+++ src/common/enums/sms-provider-functionality.enum.ts
@@ -1,0 +1,4 @@
+export enum SmsProviderFunctionalityEnum {
+  Send = 'send',
+  Lookup = 'lookup',
+}